trRamStartLow and trRamLimitLow are not optional
There are still some contradictions between trRamStartLow and trRamLimitLow being optional or not. We have agreed that these are NOT optional (which is also reflected in table 4.3.2 "Summary of Trace RAM Sink Registers"), but the two descriptions of the fields say that they may not be implemented (these sentences are probably left-overs of earlier versions of the spec, when the fields were still declared as optional).

Also, the discovery mechanism should be more precise, as "read 0 means not implemented" only applies to the two optional registers trRamStartHigh and trRamLimitHigh. It does however not apply to  trRamStartLow and trRamLimitLow.